Make fasterq dump obrigatory and add a nice exception
At first we were considering using fasterq-dump as one of various
options to download data from SRA. After a more careful consideration,
we're no longer planing on supporting any other way of downloading
the data besides fasterq-dump.

The reason is that fasterq-dump is, well, faster. It's not very
hard to install either. So it doesn't really make sense to support
slower and clumsier alternatives.

This commit also adds a nice message in case fasterq-dump is not
available, detailing the steps to install it.